Wake/Function/requirements.txt
SpeechRecognition==3.10.0
pyaudio==0.2.13
requests==2.31.0
python-dotenv==1.0.0
websocket-client==1.7.0 